refactor: reexport things from mod.ts
