+++ src/context.ts
@@ -1,5 +1,5 @@
 import { Context, NextFunction } from "./deps.deno.ts";
-import { Commands } from "./mod.ts";
+import { Commands } from "./plugin.ts";
 
 export type CommandsFlavor<C extends Context = Context> = C & {
   /**
